Release open-source full Pepper&Carrot website!
This pushes a cleaned and lightweight version of the website of Pepper&Carrot
as it is now online.

All the database was too big to be hosted here (<20GB) so I created a fake database with
colored image proxy colored and fake episodes. You can find them on 0_sources and data/images
mainly. This strategy reduced the full website to 20MB.

A full day was necessary to clean the comments and remove any sensible informations.
LICENSE: GPL V.3 for the code ( see LICENSE ) and for the artwork CC-BY D.REVOY.
License of PluXML core and PluCSS is also GPL3 ; Parsedown MIT License; and for
the plugins credits, read plugins/nameoftheplugin/info.xml file.