[Supportme refactoring, part2] aesthetic, ergonomy and cleanup
Alternatives icons were flashy and the website was looking bad:
They are now following the link color scheme of website. It is
now obvious the service I encourage user to use. But the
alternatives are still proposed and easy to discover.

The cover3 has now built-in image modification for better readibility
of text printed over it. No need of a white box in background anymore.

"More information" button is back. It targets the user confused by
all donation; "why donation?". The button links to the Philosophy
page to explain why. It was something present in old P&C website
during a long time.

Layout is now fixed for tablet and smartphone. Not 100% ideal, but it
works on the simulator proposed by Firefox. I'll do adjustment later
probably based on real smartphone and tablet observation (but I need
to upload online modifications first)

I deleted the versioning in the comment of header: now the website is
on GIT ; it is the best versionning possible. The version use the date
YYMMDD ; it is only usefull from time to time to force update of the Css
on the webbrowser cache.