Maintain "Follow P&C on" top icon and "Community" static page
Spring cleanup: the top icon colored are more and more scary;
poeple with concern about privacy but limited knowledge don't
trust this type of icons ( often provided by service who
polluate with tons of trackers and js scripts). Let's change
them for homemade chocolate colored symbolic icons.

Depreciation of G+ as a official community: the network feels
dead to me, and for those with not a lot of concern about privacy
I already maintain Twitter and Facebook. I think I cover already
a lot the non privacy world.

Removal of "Webcomic" (it had only Pepper&Carrot Mini) as a
category: this is a derivation and require a blog post about it
to be also a item I can search on the blog. So, a TODO...

Adding Patreon top colored icon; because Patreon is a social
network and the platform that link the most of P&C contributors
and supporters to the project. The interaction with the project
is important, vibrant as a community and it must be visible.

Restructure the "Press" and add pills about community on the web
dedicated to Pepper&Carrot. There is certainly mainy more.

Restructure the Community page to get intro/outro text, as it was
on the first version of the website with translation

Depreciate the "counter" on social network; yes, it looks like
20K follower are around Pepper&Carrot , but in fact it is hard
to see interaction with more than 200 humans (even if someday
the website knows 1500visit/min!). Also, counter are difficult
to maintain and no need to brag about this stat.

Add an illustration from sources/misc to keep same style than in
Philosophy page.