[Follow icons] (on top) Remove Patreon colored icon
Even if Patreon is now like a social network, and even if top-right
is a good place to attract eyes of reader; I prefer to put the icon
offline: the website looks better without it.
Informing about the support/donation system must be targeted in other
place; or in a more subtle way if it change the header, imo.