[refactor] Share buttons at end of blog-post and pages
The new buttons looks more like actionable buttons than simple link
but they don't look like javascript tracker icon of the network; I
hope it feels safe, because it is: they are just html links themed.
No spying possible here.

The default color is now bright grey to not steal focus of the content
and tiny icon "share" like on smartphone decorate the button to ease
understanding of the action.