[cache] Remove broken gzip output instead of emulating it
PluXml's gzip support should be duplicated in the cache,
but instead I'm removing it because it's broken.
In actual production Apache uses mod_deflate to gzip.

It's broken because PluXml is adding newlines before the
gzip-encoded output, and this upsets Firefox badly.
A new line is added when the PHPSESSID cookie is set.
Another is added when the plxMyMultiLingue cookie is set.

It's probably due to the Pepper&Carrot theme not clearing
the output buffer and files having spaces after the '?>' tag.