[cache] Emulate language handling properly
This adds proper language handling to the cache.

The default language oddity mentioned in the commit can be seen
in plugins/plxMyMultiLingue/plxMyMultiLingue.php in its constructor:

It calls parent::__construct($this->lang);
instead of parent::__construct($default_lang);
This sets $this->default_lang to $this->lang,
effectively makes validLang always return $this->lang

I'm unsure if which part of this is a bug since the purpose of the code
is to change how PluXml handles multiple languages.
In spite of all this somehow the default language still becomes
English, the cookies just seem to be broken.