[new-feature] showVignettePlus: display a rich formated artwork
Thanks to how consistent the 0_sources of P&C is; I created a new
template that deduce the link of the hi-res, the sources zip,
image size, link to the license.

It will help to post artworks; doing this type of footer under the
picture manually was really long, painful and was a reason why not
a lot of sketches or artwork were posted on the blog while I post
them more often on social network where it is easier to fill just
a picture + a description.

For sure, PluXML is backend UI is bloated for my use and doesn't
age well. It's like having the cockpit of a rocket to drive a
bycicle to buy bred on next street...

This commit tries to smoothen this experiences.