[plugin][plxMyAkismet] Upgrade compatibility to PluXML 5.5
This plugin was half working: the spam were (partially) well
detected. But it was impossible to delete them without editing
them one by one. Something I did, 5 time a day during around
three year. This morning; I decided to have a look at how I could
upgrade the plugin (not maintained since 2005 here:
https://github.com/Pluxopolis/plxMyAkismet ).

I found it was just variable *renamed*. This question me about the
PluXML project breaking compability of plugins just for the
aesthetic of a variable name. It sounds like breaking a machine just
for naming a knob; the thing the user will never see. There is
certainly a lesson to learn from that and something I'll keep in mind
for the management of Pepper&Carrot.