Fix homepage adress like https://www.peppercarrot.com/fr/
This type of URL was permitted before, but wasn't working: the
language in the thumbnails of the episode was rarely loaded and
resulted in a partial translation of the page.

Also, the pills on the top to launch a language were not proposing
this type of URL.

This type of URL helps at sharing the project with a target lang and
launch immediately the translated version; and the adress looks also
good this way.

Note: needs test; might injects weird behavior and be reverted.