Remove all line-breaks on Philosophy translation.
Line break was introduced when the website wasn't responsive
and was a nice way to keep the paragraph not long and manage
a "one sentence per line style".

But we are in 2019 now ; let the line flow behind each other in
paragraph and reflow depending the device and user's zoom.