Add a translation key for the main title.
Now it is dynamic; we can propose translators to write their own
translation for the title; this was a request from @arildtorvundolsen
by email this morning and I find it a very good idea. The [nn]
translation is right now the only one to use it; other still have
'Pepper&amp;Carrot' and it will be translator choices to migrate
to a fully translated title or not.