Display the number of comments on the footer of episodes
This change also now retrieve dynamically the URL on davidrevoy.com
to add a comment. This is done thanks to my blog sharing a JSON file
with all the information. The file is cached at the root of 0_sources
for performance reasons: parsing json with PHP from a remote website
was very slow.

TODO: check how the system could work in case I post a new episode
and the comments.json files is not upgraded correctly. From what
I expect; the link shouldn't display (empty <a> href) and the number
shouldn't appear. That's ok :) The blog post follow often posting a
new episode in the 5 or 10 minutes after.

Note: this commit depreciate the built-in manual listing of URL
case by case.