Switch background-color of episode thanks to data in info.json
info.json is a new file at the root of episode; meant to replace
info.yaml in long term. This file will contain specific metadata
about the episode.

A new one is the background-color Hex color; this one allows to
decorate the background of the website to get along with the color
of the panels.

It will be used with the dark episode 31.