Fix the "Translation" page in Sources
This one was framed (not fully visible) and the date were all
showing the same.

For the date, it was because the system was based on the date of the
cover jpg in the low-res folder. The date was all the same because
I pushed recently a rerender.

For the frame, its because of a theming tweak I made for the new
XYZ page and propagated this change blindly to all subsources
pages.

I switched the date system to scan the SVG source of the cover,
and removed the div header.