[sources] Merge category "original" and "inks" into a new one "sketchbook"
I'm working right now on the book project; the book 4 is an artbook.
I need to collect a lot of sketch and this categories required a cleaning.
Now the page looks like "misc" or "artworks", easier to manage.

Also, I kept the keyword "original" and "inks" to be redirected to
sketchbook, because I don't want to break any URL or bookmark in the
wild.

Sketchbook already has 100 artworks in it. 7GB of happiness.