Hide XYZ (new-ep) episode during writting/alpha step
I received feedback about it, while many part are in refactoring.
When things are not ready for public; better to hide them.
To hide a comic episode from the source; a simple "secret" empty
file at the root of the episode will hide the access to see the
source of the episode.

TODO: add this file by default in the episode creator on tools
(after renderfarm refactoring)