Fix #56: Direction issue while printing the label of RTL languages
This fix implement a dir="auto" method to all the field of language
selection. It let the browser figure what the best direction to print
the text. It works on Firefox.

Drawback: it adds 10 characters to the HTML per languages; and that's
500 characters for 50 lang... a bit heavy to fix two labels only and
also knowing that the number of rtl lang is limited... but with a
limited time budget, this is the easiest fix.

Maybe for later:
1. build a list of lang and apply "auto" to only known rtl langs.
https://stackoverflow.com/questions/13781600/php-detect-text-direction-ltr-rtl-from-ua-string-or-external-service-is-it
the list is not long: http://www.i18nguy.com/temp/rtl.html