[G1] Remove invitation to support Pepper&Carrot via payment using G1
The link was dead (I noticed thanks to a contributor giving me feedback about it),
and I'm spending more time to maintain this than get any benefit from it:

- Adding credits to Pepper&Carrot based on G1 donation was extra work.
- After years, I unfortunately never found a way to spend my G1; I have 18K... frustrating
  especially in difficult periods or when I'm facing big invoices or taxes...
- The certification require maintainance yearly, not a lot but still more than just 0.

For all these reasons, I think it is better to delete this possibility than giving the illusion
to the audience they can help me with G1 donation.

It was an interesting experiment, I'll keep my account in case I can use it later and wish
to this beautiful project a lot of good things for the future.